Allow Pootle to search for fields.
This is the codebase for implementing search for fields (GUI follows).
Important note: this commit introduces changes in terms of what information and under which names is stored by the indexing engine, so from now on it is necessary to rebuild the search indexes in order to work this functionality.
Specifically, "msgid" and "msgstr" fields now are known as "source" and "target". Additionally, "notes" and "locations" fields have been added.

svn path=/src/trunk/Pootle/; revision=8050