Replaced the old lazy evaluation code with the code in util.
Where the old code used a dictionary, the new code uses
attributes.

There is a new property: has_suggestion. This uses the
new method combine_file_failures in statistics.py to
quickly determine if a file has suggestions. This
replaces a previous check which read all the stats
to determine whether the "hassuggestion" test was
present.

svn path=/src/trunk/Pootle/; revision=8252