Here we provide Django ORM implementations of the SQLAlchemy models.
In addition to the SQLAlchemy models, there is a model called
PootleProfile. This is used to tie application-specific user
data to the generic Django User model (which contains basic
information such as the username, password, e-mail address,
superuser status).

Note the use of managers (LanguageManager, ProjectManager,
SubmissionManager, SuggestionManager) to implement queries
which cannot be expressed in terms of Django's ORM primitives.

svn path=/src/trunk/Pootle/; revision=9207