Factored out the form handling code which is common to the language and project admin pages.
Modified the projects admin template to use our formset.

Note that Django didn't like it when treestyle could be None or '',
since it had problems validating forms automatically (it thought that
treestyle's value changed, when it had, in fact, not). So let's just
make the default value 'auto'.

Also updated initdb.py to use 'auto' for the treestyle values.

svn path=/src/trunk/Pootle/; revision=9689