N.B. Each user in our system is represented by TWO objects: a User object and a PootleProfile object. The two are associated with each other.
In the old days, make_pootle_user would do the work to create both
these objects. But this was error-prone and caused some bugs that
I still don't understand.

A better mechanism is to ensure that a User object has an associated
PootleProfile by using the post_save hook of the User object.

This simplifies the code, since we can directly create User objects
now and be safe in the knowledge that associated PootleProfile
objects will always be created.

Another simplification was to get rid of save_user. This function
called the save method on both a User object and its associated
PootleProfile. But really, code that modifies either should know
which to save.

Finally, we also associate the PootleProfile of the 'nobody' user
with anonymous users.

svn path=/src/trunk/Pootle/; revision=10031