This commit changes the way alternative source language translations are retrieved.
Until now, calls to ugettext were used, which looked over all the files in a project. This is not ideal for big projects specially. The new implementation looks for the translation in the same file as the original text comes from and the matching is done using the getid() method of the unit.

This is O(n) though, and for future improvements we should think about using an index (makeindex() in translate::storage::base).

Apart from that, this commit also adds support for plural forms, which wasn't present yet.

This fixes bug 605.

svn path=/src/trunk/Pootle/; revision=10260