dear django please stop sucking.
django defaults to storing uploads in memory (need to think if that's a good idea) but the in memory file object is radically different from the real file object, and factory.get_object doesn't like the difference.

and no, just because StringIO can do it perfectly doesn't mean that django needs to implement closed for in memory files.

svn path=/src/trunk/Pootle/; revision=12690