Properly check for profile permissions.
An store string was being passed instead of an instance of the
Directory model.

Also, this now raises another problem: anonymous and default users may
have problems when submitting translations/suggestions for the first
time since we have the captcha middleware. This needs to be adapted to
understand xhr requests and act accordingly.

svn path=/src/trunk/Pootle/; revision=15900