Manage assets using webassets.
This requires the 'webassets' and 'cssmin' python packages. The first
one eases the management of assets within django, and the second one is
required to minify CSS files.

Different filters need to be used to create the bundles. For JavaScript we
use the builtin 'rjsmin' filter, but YUI or Google Closure Compiler can
be used by providing proper configuration settings.

This fixes bug 2112.

svn path=/src/trunk/Pootle/; revision=18036