Added an 'apply_markup_filter' function.
This will allow to use 'textile', 'markdown', or 'restructuredtext' in
specific textareas. It'll be an optional component.

A two-tuple 'MARKUP_FILTER' setting controls the filter used. The first
argument is the filter to be used (to disable any filters, it can be
None), and the second argument is a dictionary of keyword arguments to
be passed to the markup function.

Borrowed from http://djangosnippets.org/snippets/104/