Added the ability to define 'legal' pages.
These kind of pages are useful for Terms of Service, Privacy Policy,
License and similar information.

Pages can point to an external URL (privacy policy hosted in another
server) or provide own content.
Contents can be formatted according to the markup defined in the
server settings.

For now links are displayed in the footer, but selected pages will be
displayed on registration as well.

This is a first step towards fixing bug 2134.