Management commands for VCS actions.
This allows version control actions from the command line and allows
these actions to be performed across translation projects - at the
project level or even for the whole server. This is all still performed
on a file-per-file basis.

This is largely based on the work of Stuart Prescott, and fixes bug
1491.