Make commenting an explicit action.
Now it's possible to add new comments without submitting the whole unit.
Technically-speaking this is not true, since the unit is updated behind
the scenes. Starting from Django 1.5 we could use the 'update_fields=[]'
argument passed to the save() method to just trigger an SQL UPDATE of
the fields that we are updating.

Notes:

  - No changes have been made to term_edit.html; we should be inheriting
    the common parts from edit.html.
  - Some UI issues will be addressed as part of bug 2283.

This fixes bug 2296.