Fix shortcuts for zooming
The old ones had clashes with other navigation shortcuts.
The new ones align mostly with what was proposed.
I changed * to 0 for the zoom reset, seemed more logical.
Unfortunately - seems to not work, needs validation on other
platforms.  But I'd rather see this fixed in shortcuts.js then
try to change it.