Don't coerce u'0' to False.
This was causing to treat checkboxes with `value="0"` to be treated as having no
value at all. Since we store the value of the unit state in the checkboxes (it
can be any of UNTRANSLATED, FUZZY, or TRANSLATED), we need to treat them as
special values.

The default `BooleanField` interpreted u'0' as the checkbox being unset, so
submitting any UNTRANSLATED units (their value is u'0') would result in no state
changes at all.

Fixes bug 2620.