Split updatedb functionality.
Originally updatedb served as a command for two different purposes:
- Performing database schema upgrades.
- Performing actions specific to Pootle/TT version upgrades

Since Pootle is transitioning to South, we need to ensure that:
- Old users will still be able to upgrade
- Version-specific upgrade actions will be able to be performed once
  database schema migrations are handled entirely via South

The updatedb management command has been slimmed down to perform only
schema upgrades and only up to version 2.5. That way we ensure old users
can upgrade, and we move the version-specific actions to a new command,
upgrade. The new upgrade command will keep track of which build version of
the code is running the site in order to perform the necessary post schema
upgrade actions.

Schema upgrades starting from 2.5+ will be entirely handled via South's
migrate command.

Part of bug 2773.