Fix bug 2870
When making a search query, we first split the search string into
individual words, but for some reason if the resulting string contains
on the stopwords the searcher fails to find any results. Making a query
of the whole search string seems to fix then since PyLucene will
tokenize string itself and discard the stopwords. The resulting queries
still match substrings, reordered strings etc. so no regression is
expected.