Fix sub-list reStructuredText formatting
When using sub-lists, blank lines are required between list items;
if omitted, sub-lists are treated as normal flowed paragraphs with
embedded 1. or * in the text.

(Blank lines between list items are always allowed, and are usually
a good idea, but are optional if there are no sub-lists or embedded
un-numbered or -bulleted paragraphs between the list items.)