Make the sub-lists Markdown-compatible
This isn't strictly necessary, since these are reStructuredText docs, but
you never know when somebody might cut & paste your useful advice for use
in another context, where Markdown formatting is more likely.

For Markdown to recognize a sub-list, it has to be indented by 4 to 7
spaces.  Just indenting the sub-list items would work for Markdown, but in
reStructuredText, all the main list items will become bold and you will be
*very* confused.  (Indenting the sub-list items more than the main list text
makes them look like definition lists, where the "defined word" is bolded.)
So you need to indent the sub-list items by 4, and then add extra space
between the bullet or number and the text in the main list items that have
sub-lists, so that the main list text is aligned with the sub-list numbers
or bullets.  Now everybody is happy.

And you thought these plain-text formatting/markup languages were simple?