Decouple editor from translation projects
This is the biggest piece chunk of work to decouple the editor from
translation projects (bug 3013). There are still fixes and more cleanups
pending, but this should give a good base to start working on
cross-language/-project scenarios such as bug 2719.

The indexing part has been commented out for now, and that should be
addressed as part of bug 3028.