Black magic that fixes #4
If we retrieve the actual SQL query then the Paginator breaks for
querysets containing `.extra()` calls. This DIRTY HACK allows the actual
`SELECT SUM (*) FROM ...` SQL query to select and join the necessary
tables.

(I suspect that the current filtering constraints aren't fully populated
when the paginator tries to calculate the count, and this hack somehow
forces it.)

This might be a bug (or undocumented behavior) in Django's ORM querying
facility internals, so any further investigations should be done in that
direction.