Fix Google Translate button initialization
Notes: at MT init time, the targetLanguage member seem to be missing; the target language is now taken on editor "ready" event, when it is gathered from <textarea lang="..."> attribute (there should be a cleaner way to store and retrieve the source and target languages than it is now). Also: was there a reason to store the 'targetLang' as a propery of the plugin? I removed that code. Especially now that the target language can be different across units.