Use contact form to report string errors
The new form serves both as a way to contact site managers as well as to
report errors with strings from the editor.

This also renders useless the need for a `report_target` field in
`Project` models, so that field has been deleted too, along with its
related functionality.

This, and some other previous commits, fixes evernote#15.

Partial pick of evernote@a76f6577. The rest of the original commit is in
several previous commits.