Rewritten contact form
The new form serves both as a way to contact site managers as well as to
report errors with strings from the editor.

Under the hood, django-contact-form 1.0 has been used, thus removing the
old copy bundled in the codebase, and its behavior has been customized in
the `evernote_contact` app.

This also renders useless the need for a `report_target` field in
`Project` models, so that field has been deleted too, along with its
related functionality.

Fixes #15.