Improved JavaScript handling of the captcha
Less hackish, less requests, less duplicated code, cleaner separation of
code and not tightly coupled to the editor.

JavaScript code that makes XHR POST requests will need to pass two special
parameters (`sfn` and `efn`) which are the success and error callback
function names which will be executed once the captcha has been
successfully resolved.

Note that the captcha can now be identified via the HTTP 402 status code
(Payment Required) — the captcha can be considered like a sort of payment
the user needs to make in order to continue further. YouTube also makes
use of the HTTP 402 status code for this purpose.