Revert "Prevent Firefox from attempting to install the XPI"
This reverts commit 1d39a08282ef5385affc431810372bf66cdfa2ee.

This fixes serving files from extension actions. I didn’t understand
what I was doing back then; the redirect() is needed because it is a
special URL handled in pootle/urls.py to serve the files with
django.views.static.serve, so using normal HttpResponse() meant I was
sending the browser a text string not the actual file!

The issue of Firefox trying to install the XPI file instead of
downloading it remains unsolved.