Added missing submission type attribute
Bulk actions performed offline were missing a value for this attribute.
We're adding it for consistency and completeness. Includes data migration
code.

Pick note: When this commit was picked from Evernote fork it suffered some
changes, like replacing the migration with code in the upgrade system, and
thus increasing the build number.