Stop using `post_syncdb` signal
This hook stopped working for us since the migration to South;
previously it was run once after `syncdb` command, but now it is run
after each app migration so not all the tables we need will be created
before the hook is called.

This commit moves this code to the `initdb` command that is run after
`migrate`, with this we can now successfully install master from scratch
again.