Request specific unit ids instead of pages
The concept of pages is problematic for the editor since we are dealing
with a result set of a dynamic nature that is being changed in the server.
This resulted in unexpected behaviors and reliability was missing when
performing certain actions in the editor.

From now on the client will request an explicit list of unit ids the
server needs to query and return. The client keeps a sorted list of
unit ids fetched in the very first request, therefore it knows which units
it has and which ones are missing.

Fixes issue evernote/pootle#48.