Added 'announcement' page type
This is a hackish way to create a special regular page, and most of its
singularities are specified in view code. Announcements are not an entity
of their own and they are represented as static pages in the model and
database level. Not the cleanest and fanciest way of doing this, but since
these special pages are a temporal workaround, it's a trade-off we can
make.

Part of #59.