Drop django-registration app from Pootle codebase
Seems that the code was not on Pootle codebase because it was necessary to
localize it, but perhaps because this way was easier to release Pootle.

The new registration app doesn't save the RegistrationForm, so we can't
rely on it for saving the agreements accepted when the user registered.

Use pinned version for the package in requirements to avoid ugly surprises.

Fixes bug #2579