Profiles: Apply PR to get rid of function-based generic views
The PR is in https://bitbucket.org/ubernostrum/django-profiles/pull-request/6/

There is another PR with a slightly different approach, but when and if
django-profiles finally supports class-based generic views the code will be
quite similar to any of these two PRs.

Whitespace changes are not merged to make it easier to see the real changes.

This commit and the previous two ones fix bug #2517