Added the ability to delete the latest comment
In this context removing simply means storing an empty string for the
`translator_comment` field. Since technically speaking, removing a comment
is the same as submitting a comment, a new submission is recorded for this
action too, but it's presented differently to the user.

Likewise, the latest comment is now hidden from the results of the
timeline, thus avoiding showing a comment twice.

Fixes #88.