Remove unused code on util
Followup for previous commit.

After some changes required to use reverse to get the URLs, the link
parameter received by the util.edit() is no longer a string with string
placeholders (%s) but only a function that accepts the current form
instance and returns a valid URL for that instance.

This commit removes the code that handled the "link is a string with
string placeholders" case.