cache: Cache the top stats for a much longer time
Previous behaviour had us cache for x*POOTLE_CACHE_MIDDLEWARE_SECONDS
which is 10-15 minutes by default. That is ridiculously quick for
something that is very expensive to compute and doesn't change that
much.
By default we now cache for a whole day and this can be changed in the
new setting TOP_STATS_CACHE_TIMEOUT.