docs: older pip might install pre-releases
Looks like pip might install e.g. 2.5.1-rc1 when 2.5.1 has been
released.  It would seem that their version detective work isn't working
in those cases.

This occured on Debian Wheezy (7.4) and pip 1.1

I don't think we want to remove the old versions from pip because we
want to allow fallback path for users if required and because these are
FOSS releases so we shouldn't just remove them.

So make a note in the install guide for this possible occurance.

Fixes bug 3185 to some extent.