Require non-admin users to always clear the fuzzy flag
Non-admin users now have the fuzzy checkbox disabled by default. It'll be
shown if the initial state of the unit is fuzzy or some other action (MT
translation, auto-translation etc.) turned the unit into fuzzy state.

Whenever the fuzzy checkbox is displayed for non-admins, they are forced
to clear it, otherwise the submit button won't be enabled.

This covers the frontend side of evernote/pootle#157.