Followup for .pending and .tm removal
There was some references to the code dropped in the previous commit, so we
removed them. But since they were on the upgrade code, removing part of that
code rendered it useless, so we dropped the upgrade path for the ancient
Pootle versions (2.1.0 and older).

So to avoid surprises we added a check in the setup command to warn people
that they are trying to upgrade from a really old version and ask them to
do an intermediate upgrade.

Added a note about this on the upgrade instructions and included references
in the release notes as well.