models: Drop natural key handling everywhere.
This is a partial revert of the very old commit fc0b5f8e, which was done
to improve dumpdata and loaddata commands at a time when they were still
useful.
This is no longer the case since those commands shouldn't be used,
especially not to convert between databases.

Also update notes on database migration

Conflicts:
	docs/releases/2.5.2-rc1.rst
	pootle/apps/pootle_app/models/permissions.py
	pootle/apps/pootle_language/models.py
	pootle/apps/pootle_profile/models.py
	pootle/apps/pootle_project/models.py
	pootle/apps/pootle_store/models.py
	pootle/apps/pootle_translationproject/models.py