Editor: auto-approve the exact suggestion
When the text being submitted matches one of the existing suggestions, the
suggestion will be accepted and the next unit will be loaded.

Note that the implementation is completely lame, ugly and fragile, prone
to be broken in the future. Suggestions need to be models on their own,
and we need to move away from tying model functionality from the DOM.

Fixes #160.