pep8: silence new E265 for block comments
pep8 v1.5.0 introduces E265 block comment should start with '# '.  Since
we blanked selected E26 we got E265 as soon as it landed.  We have too
many errors in Pootle related to this, mostly where code has been
commented out.  Some should be cleaned up but there is a conflict churn
issues that we'll hit with this.