Django 1.5: killed djblets
This old bundled dependency contained code incompatible with Django 1.5,
but apart from that the feature the code was depending on (keeping track
of site's title and build versions) is of no use for us anymore.

The commit moves the site's title straight away into the settings, so it
should be configured right there.

On the other hand the upgrade machinery has been taken out, as it's of no
pratical use for us. Our releases just depend on DB schema and data
migrations.

Fixes #177.