User: removed calls to `get_profile()`
This was both a method of old `User` models (which returned the linked
`PootleProfile` instance) as well as a function which returned the special
`nobody` user (which references any anonymous user).

The former is useless now, since there's no distinction between profile
and user models. The later has been converted into a class method of the
`User` model.

Part of #199.