Recover UI: Display instructions in sidebar
This is a temporary recover until we manage to get this blended into
staticpages.

This includes the ability to edit in the sidebar the instructions for
the current language, translation project or project being displayed.
If no instructions exist it is possible to create them using the same
editing form in the sidebar.

If the overview is for a translation project then the instructions for
the language and the project are also shown (if they have any
instructions), but it is not possible to edit them.

Includes extensive changes to support editing two forms at the same in
the sidebar, because it is now possible to have a form for editing the
instructions and another one for editing a goal.