Bugfix: Display all failing checks in editor dropdown
This is because of 8a0718d0, that is necessary because in the stats
we only need critical checks, but in the editor dropdown we need all
the checks, including cosmetic and functional. This was no problem
for Evernote because they only have critical checks.

This is a dirty hack that needs to be partially undone in a future not
yet created followup commit that detaches get_error_unit_count() from
get_checks(). See the TODO note.