Display failing checks below the roleout
Not sure why there were being hidden. Despite the sum of critical counts
doesn't match the one in the translate action, or the sum of critical
counts in the stats table that is because here we show the number of
critical errors, and in the other two places we instead show the number
of units with critical errors (it is possible for a unit to have several
critical errors).