Unit: forbid overwriting source texts from the web interface
Overwriting source texts in this context is undesired, which ended up in
mangled DB content because of evernote/pootle#278.

The reasoning behind allowing to edit source texts can be explained with
the fact that terminology units have been sharing the editing/validating
form with regular units. It's not the case anymore — these are very
different contexts with different use-cases.