Admin: fixed overextend for language and project pages
This was being caused by the view code that was populating the template
context. It turns out that because of a newbie Python mistake (never, ever
use mutable types as default function arguments), the template context was
being shared across requests, and every time a view using `util.edit()`
was hit, a path was being removed from the search path. That's why it
worked on the first couple of requests but failed in the next ones.

The reason why this was only happening for language and project pages is
simple. Any other views returning the buggy `utils.edit()` were at least
defining a default context, therefore a copy of the provided context
dictionary was being used on every call and not a "default" shared one.

Fixes #299.