Make translated words totals equal in differnet parts of report
When we summed translated word count by translation project we could get
a total count that differs from total count in summary. That is a side
effect to see the whole numbers for a translation project.
Totals must be equal. And subtotals are calculated for that total in
whole numbers.