Travis: install lxml without compile optimisations
Use CFLAGS="-O0" as recommended by
http://lxml.de/installation.html#installation to speedup installation
times in Travis builds.

This is applied to all installations here, should maybe be limited to
lxml.

The next step wpuld be to create precompiled wheel packages to speedup
setup.

Without this change we take on average 2,5min to setup the packages for
testing. After this we shave off 1,3min i.e. setup takes on average 57s