Avoid clashes when creating announcements for project descriptions
If by chance some previous announcement existed for a given project,
when migration its description to an announcement the new announcement
can't be created. This is solved by concatenating the description to
the existing migration.

Also alter other migrations for moving descriptions to depend on the
migration (added in previous commit) to alter the virtual paths for
existing announcements.