Store: use custom file system storage class
Subclassing allows us to customize the `location` without interfering with
migrations – they use serialized Python code and the subclass allows to
hide the actual location value from them.

With the previous implementation the developer's `settings.PODIRECTORY`
would be leaked, and other developers would see a changing field (and
therefore a warning from Django) all the time.

This commit also drops the `upload_to='fish'` keyword argument, which was a
dummy placeholder because it was a required parameter, but it becomes
optional in Django 1.7 and it's safe to remove it.

Fixes #3557.