Revision: switched counter to Redis
The DB model has been replaced by a thin wrapper around the caching API of
Django, which in turn we are forcing to be a Redis backend, thus enabling
atomic counters.

This also requires switching tests to use the `LocMemCache` backend,
because the operations that require the revision to be incremented will
fail otherwise — the `DummyCache` is a no-op.

If for some external reason the revision counter has been lost, it's
possible to restore it via the `revision --restore` command.

Fixes #3364.