Lightbox: render it as a layered component
The previous implementation had big issues because the lightbox was
displayed in the context where the rendering happened, and it created nasty
layouts which were prone to be altered by browsers due to unsupported
element nesting (e.g. a table within a link). Styles could be an issue
too, because they are inherited from the current context as well.

The natural thing here is displaying lightboxes in isolated contexts or
layers, appended to the body of the document.

The only thing left from the previous implementation is handling of
tabbing focus, which always needs to remain within the boundaries of the
current lightbox.