localtm: fix ES based TM to do proper Levenshtein matching
Seems that with the current setting we're just getting match score and
not actually doing Levenshtein matching.

MIN_SCORE was suggested to be set to 0.1, yet score values returned by
match are more likley to be >1 so almost anything matches.

Adding 'fuzziness' to the ES search gets us Levenshtein results.  So we
repurpose MIN_SCORE to provide ES with the 'fuzziness' setting. These
can take on a number of values.  We default to 'AUTO' which adjusts the
permissible edits based on query length.  This should give us pretty
good results for short and long strings.