Cache Store.serialize() results
Also adds a new 'exports' cache.

NOTE:
Having exports cache on /var/cache/ requires creating dirs and
assigning permissions.

It is better to store the cache on a directory close that we can
access easily and that we can ensure already exists when deploying.

Also spread the settings to all the places where CACHES is defined,
and include the necessary excludes on gitignore.