Cleanup: decoupled search from the editor
At the very least now the editor doesn't update the search UI as it did
before. Unfortunately the state that comes via URL params still needs to
be read in the editor, thus it also manipulates the search state.

So while some ugly things are still going on, this can be considered as a
first step towards the componentization of the module.