Add command to regenerate the quality checks descriptions on demand
The resulting descriptions are saved as the help/_ttk_quality_checks.html
template. That filepath is added to ignore in order to avoid inadvertently
commits changes in this template.

This requires the latest available code of Translate Toolkit in order
to work as expected.

The docutils library is also required. If not present an error message is
displayed asking for it.

Fixes #3604.