Browser: use valid CSS ids for stats
Invalid selectors, along with a mismatch in how these were being handled,
prevented stats from some items (those with codes that contained `.` or
`@`) to be displayed.

Consider this a temporal workaround — the preferrable way is to avoid
using the DOM as a data store, which avoids issues of this nature
altogether.

Fixes #3729.