docs: cleanup hacking guide
This guide is schitzo.  Its so chatty, but I'd hope someone wanting to
develop at least knows some of the things that we're teaching them here
already.  Its a complete repeat of what I assume we have in the
installation guide.  So its much info repeated that goes out of sync now
in two places.

I didn't try fix that, just tried to improve what we have here.