ribbon: add ribbon to master template
Mostly nicked by looking at the Evernote staging instances.  Logic being
that if POOTLE_INSTANCE_ID is set we will display the ribbon, otherwise
not.

The URL to Github is still hardcoded where really it should point to the
remote that is actually being used by the checkout.  In most cases that
would be translate/pootle so it should be OK in the general case.  It
sems that `git ls-remote --get-url' would be able to do that.