Staticpages: removed implicit auto date behavior
On the one hand, there's already `default=now`, which is exactly the same
as `auto_now_add=True`.

On the other hand, this was preventing tests from overriding
`LegalPage.modified_on` and `Agreement.agreed_on`, as the built-in
behavior is implemented via a `pre_save` signal that messes things around
— we better be explicit whenever we want to automatically update
timestamps.

Picked from 7a691358.