staticpages: seems we never needed this migration
Blame Jerome and Dwayne if it is in fact needed.

user_id has never been a link to PootleProfile and has always been a
link to User.  So this migration is copying an incorrect PootleProfile
ID into a valid User ID.

This has seemed to work because there is an almost 1to1 mapping between
Users and PootleProfile.  And it just never caught us until now.  Now
that we ended up with some non-unique combinations we threw and error.

The difficulty in discovering this was that the convoluted JOINs made us
focus on what the SQL was doing.  It was only when we stepped back and
asked why are we doing this enough that we wondered if this situation
was true and if this migration was even needed.

So essentially we had user who potentially hadn't signed an agreement
marked as having agreed and visa versa.

So yip, it was never needed.