version: without path this doesn't work
Seems that in a server env we don't have 'git' in the path so we need to
specify the path.  Its a pain since this could be located in
'/usr/local/bin'.  But making it configurable seemed like overkill.  So
lets do it like this until the use case actually emerges and we need to
go setup some sort of configuration setting.