localtm: use bulk upload in update_tmserver
This implements the bulk upload mechanism of elasticsearch.  An iterator
is supplied to the bulk helper which by default does chunks of 500
units which we craft from the QuerySet iterator.

This leads to a 10X speedup in the upload.

This code should really be moved into the actual TM ES backend.  But it
was a quick workaround for LibreOffice deployment that is useful as it
is now.  It would probably also love some loving refactoring.