Drop pinned requirements
This was used in the release process for testing the latest versions
of the requirements, but since we are not setting pinned requirements
when creating the package there still exists the possibility that a
requirement publishes a new problematic release.

This also confused users in the past.

Reverts b2441b83 and 85d77b43.