TranslationProject: introduced `disabled` property
This was previously available as a field on its own but it was removed as part
of #3674. However the disabled state can be computed according to the
obsoleteness of the TP's directory and the project's disabled state. This comes
handy as it fixes a regression where disabled TPs could be displayed differently
in the browsing pages.

Fixes #3944.