TimeSince: implemented graceful auto-update
Now this component is practically independent: it doesn't rely on an external
party querying by class name and performing a refresh on a set interval.

Also improves the logic to auto-update, as this is now dynamic, and changes
according to the time delta between now and a given past time: every 30 seconds
up to the minute, every minute up to the hour, every hour up to the day, once a
day otherwise.