Revert "Cache: simplified invalidation of project resources"
This reverts commit 3b6100b3374a28c7572a7aba790e9296bafe9170.

It'd be necessary to monkeypatch the caching backend module to add the
`delete_pattern()` method, but there's no reliable way to do so just yet
(pytest#746). So it's better to keep some ugly code around rather than having
broken tests.

Also it's worth noting that we switched to `LocMemCache` for testing the
revision module, although in general we always want to use the `DummyCache`
module during testing. However switching caching subsystems during run-time is
not a matter of altering settings, since the cache singleton is assigned at
import time.