Stats: remember expanded state when returning from the editor
User interactions with the expanded stats widget will add or remove a `?details`
query string to the URL, which then will be taken into account when creating the
URL to redirect users back to the page where they came from.

When this query string is present on initial page load, the widget will be
automatically displayed expanded. The previous behavior of displaying the
expanded state for stores is kept intact.

Fixes #3784.