Editor: ensure default properties are updated early
This code has to be run early on, at least before it checks a suggestion with
the same value as being submitted exists. In such case, it'd accept the
suggestion, but since the default properties were not updated, it'd think the
unit was still dirty, hence triggering the confirmation popup.

Fixes #4076.