Export view: fixed regression for admin users
The `.count()` call is ignoring the columns in the `select_related()` call which
are present in the `UnitManager`, as Django can't know if it's the right choice
for it or not. This commit works around that by calling `.annotate()` on the
affected queryset.

As a side note, we are probably better off moving away from Django ORM's
`.extra()` APIs, as they're troublesome and will be deprecated at some point in
the future.

Refs. https://code.djangoproject.com/ticket/11459#comment:9

Fixes #4088.