Auth UI: added scaffolding for actions and reducers
This is a initial step towards porting the UI away from Flummox to Redux.

In this commit the skeleton of actions from Flummox have been ported, and the
same has been done with stores, which are now reducers. This is just
scaffolding, so there's no actual functionality yet, and it hasn't been glued to
the UI either. The Redux library itself has been added as well.