Auth UI: switched app's entry point to react-redux
The glue between the store and the UI is managed via the `<Provider />`
component, which wraps the "smart" `<AuthController />` container component.
This component, in turn, is wrapped in a `connect()` call, which signals that it
wants to have access to the app's store. The `<Provider />` component is fed a
store built out of the root reducer; this supersedes the flux object from
Flummox.

Note the identity function is used as a selector in the container, as it's not
necessary to perform any filtering on the store for the time being. Also note
the `<FluxComponent />` is superseded by react-redux's `<Provider />`