Editor: fix retrieving results via amaGama
The problem here was that there was a strict equality check in place (`===`),
and `uid` and `unit.id` were of different types, so it always returned `false`.
In any case, the uid check is redundant as ongoing requests are always aborted
before initiating new ones, so removing the check is enough.