test_checks: handle all checkers, not just ENChecker
Use either ENChecker if settings is configured or use all the possible
Translate Toolkit checkers.

We don't handle the case where the test needs location, rather bailing
out if we get an AttributeError. Not ideal but good enough to make the
tool usable again for 99% of the cases.

Made source and target work of Translate Toolkit that needs unicode
strings.

Fixes #4205 part 2